added SerialConnection as interface
